// src/app/infrastructure/models/api-application-user.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ApiBox } from '../models/api-box';
import { ApiGenderEnum } from '../models/api-gender-enum';
import { ApiInterest } from '../models/api-interest';

/**
 * Класс, представляющий пользователя приложения, наследуемый от IdentityUser
 */
export interface ApiApplicationUser {

  /**
   * Поле "О себе"
   */
  aboutMe?: string | null;
  accessFailedCount?: number;

  /**
   * Список коробок
   */
  boxes?: Array<ApiBox> | null;
  concurrencyStamp?: string | null;
  email?: string | null;
  emailConfirmed?: boolean;

  /**
   * Имя пользователя
   */
  firstName?: string | null;
  gender?: ApiGenderEnum;
  id?: string | null;

  /**
   * Список интересов
   */
  interests?: Array<ApiInterest> | null;

  /**
   * Фамилия пользователя
   */
  lastName?: string | null;
  lockoutEnabled?: boolean;
  lockoutEnd?: string | null;
  normalizedEmail?: string | null;
  normalizedUserName?: string | null;
  passwordHash?: string | null;
  phoneNumber?: string | null;
  phoneNumberConfirmed?: boolean;
  securityStamp?: string | null;
  twoFactorEnabled?: boolean;
  userName?: string | null;
}